refactor(core): extract Command/ContextId to dbflux_core::keymap_types (S2) (#102)
