+++ crates/dbflux_app/Cargo.toml
@@ -53,4 +53,4 @@ influxdb = ["dep:dbflux_driver_influxdb"]
 mssql = ["dep:dbflux_driver_mssql"]
 lua = ["dbflux_lua"]
 aws = ["dbflux_aws", "dbflux_ssm"]
-mcp = ["dbflux_mcp", "dbflux_mcp_server", "dbflux_policy"]
+mcp = ["dbflux_mcp", "dbflux_mcp_server", "dbflux_policy", "dbflux_core/mcp"]
