+++ crates/dbflux_app/src/keymap/mod.rs
@@ -3,13 +3,10 @@
 //! This module contains pure domain types with no GPUI dependency.
 
 mod chord;
-mod command;
-mod context;
 mod focus;
 mod keymap_layer;
 
 pub use chord::{KeyChord, Modifiers, ParseError};
-pub use command::Command;
-pub use context::ContextId;
+pub use dbflux_core::keymap_types::{Command, ContextId};
 pub use focus::FocusTarget;
 pub use keymap_layer::{KeymapLayer, KeymapStack};
