+++ crates/dbflux_core/Cargo.toml
@@ -28,6 +28,9 @@ tokio = { version = "1", features = ["sync", "time"] }
 futures = "0.3"
 dbflux_policy.workspace = true
 
+[features]
+mcp = []
+
 [dev-dependencies]
 dbflux_test_support.workspace = true
 tokio = { version = "1", features = ["rt", "macros", "sync", "time"] }
