+++ crates/dbflux_core/src/keymap_types.rs
@@ -403,3 +403,197 @@ mod tests {
         assert_eq!(Command::SaveQuery.display_name(), "Save");
     }
 }
+
+/// Identifies the current UI context for keybinding resolution.
+///
+/// Different contexts have different keybindings. When a key is pressed,
+/// the system first looks for a binding in the current context, then
+/// falls back to the Global context if no match is found.
+#[allow(dead_code)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+pub enum ContextId {
+    /// Global context - keybindings available everywhere.
+    #[default]
+    Global,
+
+    /// Schema tree navigation in the sidebar.
+    Sidebar,
+
+    /// SQL editor area.
+    Editor,
+
+    /// Results table area.
+    Results,
+
+    /// Background tasks panel.
+    BackgroundTasks,
+
+    /// Command palette modal (captures all input).
+    CommandPalette,
+
+    /// Connection manager modal (captures all input).
+    ConnectionManager,
+
+    /// History modal (captures all input).
+    HistoryModal,
+
+    /// Any text input is focused and receiving keyboard input.
+    TextInput,
+
+    /// A dropdown menu is open and receiving keyboard navigation.
+    Dropdown,
+
+    /// SQL preview modal is open (captures all input).
+    SqlPreviewModal,
+
+    /// Context menu is open and receiving keyboard navigation.
+    ContextMenu,
+
+    /// Confirmation modal is open (dangerous query, delete, etc.).
+    ConfirmModal,
+
+    /// A navigable form is in `Navigating` mode (j/k/h/l move focus ring).
+    FormNavigation,
+
+    /// Execution context bar (Connection/Database/Schema dropdowns).
+    ContextBar,
+
+    /// Audit event viewer row list.
+    Audit,
+
+    /// Event-stream picker modal (collection child picker).
+    EventStreamsPicker,
+}
+
+impl ContextId {
+    /// Returns the parent context for fallback keybinding resolution.
+    ///
+    /// Modal contexts (CommandPalette, ConnectionManager) and input contexts
+    /// (TextInput, Dropdown) have no parent because they capture keyboard input.
+    pub fn parent(&self) -> Option<ContextId> {
+        match self {
+            ContextId::Global => None,
+            ContextId::CommandPalette => None,
+            ContextId::ConnectionManager => None,
+            ContextId::HistoryModal => None,
+            ContextId::TextInput => None,
+            ContextId::Dropdown => None,
+            ContextId::SqlPreviewModal => None,
+            ContextId::ContextMenu => None,
+            ContextId::ConfirmModal => None,
+            ContextId::FormNavigation => None,
+            ContextId::ContextBar => None,
+            ContextId::EventStreamsPicker => None,
+            ContextId::Sidebar => Some(ContextId::Global),
+            ContextId::Editor => Some(ContextId::Global),
+            ContextId::Results => Some(ContextId::Global),
+            ContextId::BackgroundTasks => Some(ContextId::Global),
+            ContextId::Audit => Some(ContextId::Global),
+        }
+    }
+
+    /// Returns true if this context captures all keyboard input (modals/inputs).
+    #[allow(dead_code)]
+    pub fn is_modal(&self) -> bool {
+        matches!(
+            self,
+            ContextId::CommandPalette
+                | ContextId::ConnectionManager
+                | ContextId::HistoryModal
+                | ContextId::TextInput
+                | ContextId::Dropdown
+                | ContextId::SqlPreviewModal
+                | ContextId::ContextMenu
+                | ContextId::ConfirmModal
+                | ContextId::FormNavigation
+                | ContextId::ContextBar
+                | ContextId::EventStreamsPicker
+        )
+    }
+
+    /// Returns true if this context is the audit viewer context.
+    pub fn is_audit(&self) -> bool {
+        matches!(self, ContextId::Audit)
+    }
+
+    /// Returns a human-readable name for this context.
+    #[allow(dead_code)]
+    pub fn display_name(&self) -> &'static str {
+        match self {
+            ContextId::Global => "Global",
+            ContextId::Sidebar => "Sidebar",
+            ContextId::Editor => "Editor",
+            ContextId::Results => "Results",
+            ContextId::BackgroundTasks => "Background Tasks",
+            ContextId::CommandPalette => "Command Palette",
+            ContextId::ConnectionManager => "Connection Manager",
+            ContextId::HistoryModal => "History",
+            ContextId::TextInput => "Text Input",
+            ContextId::Dropdown => "Dropdown",
+            ContextId::SqlPreviewModal => "SQL Preview",
+            ContextId::ContextMenu => "Context Menu",
+            ContextId::ConfirmModal => "Confirm",
+            ContextId::FormNavigation => "Form Navigation",
+            ContextId::ContextBar => "Context Bar",
+            ContextId::Audit => "Audit Viewer",
+            ContextId::EventStreamsPicker => "Event Streams Picker",
+        }
+    }
+
+    /// Returns all context variants in display order.
+    pub fn all_variants() -> &'static [ContextId] {
+        &[
+            ContextId::Global,
+            ContextId::Sidebar,
+            ContextId::Editor,
+            ContextId::Results,
+            ContextId::BackgroundTasks,
+            ContextId::CommandPalette,
+            ContextId::ConnectionManager,
+            ContextId::HistoryModal,
+            ContextId::TextInput,
+            ContextId::Dropdown,
+            ContextId::SqlPreviewModal,
+            ContextId::ContextMenu,
+            ContextId::ConfirmModal,
+            ContextId::FormNavigation,
+            ContextId::ContextBar,
+            ContextId::Audit,
+            ContextId::EventStreamsPicker,
+        ]
+    }
+
+    /// Returns the GPUUI key_context string for this context.
+    pub fn as_gpui_context(&self) -> &'static str {
+        match self {
+            ContextId::Global => "Global",
+            ContextId::Sidebar => "Sidebar",
+            ContextId::Editor => "Editor",
+            ContextId::Results => "Results",
+            ContextId::BackgroundTasks => "BackgroundTasks",
+            ContextId::CommandPalette => "CommandPalette",
+            ContextId::ConnectionManager => "ConnectionManager",
+            ContextId::HistoryModal => "HistoryModal",
+            ContextId::TextInput => "TextInput",
+            ContextId::Dropdown => "Dropdown",
+            ContextId::SqlPreviewModal => "SqlPreviewModal",
+            ContextId::ContextMenu => "ContextMenu",
+            ContextId::ConfirmModal => "ConfirmModal",
+            ContextId::FormNavigation => "FormNavigation",
+            ContextId::ContextBar => "ContextBar",
+            ContextId::Audit => "Audit",
+            ContextId::EventStreamsPicker => "EventStreamsPicker",
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::ContextId;
+
+    #[test]
+    fn history_modal_is_modal() {
+        assert!(ContextId::HistoryModal.is_modal());
+        assert_eq!(ContextId::HistoryModal.parent(), None);
+    }
+}
